requirements.txt
autopep8==1.4.3
braintree==3.52.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cssmin==0.2.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.1
Flask-SocketIO==3.3.2
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10
jsmin==2.2.2
MarkupSafe==1.1.0
monotonic==1.5
newsapi-python==0.2.3
peewee==3.1.0
peewee-db-evolve==3.7.0
peewee-validates==1.0.7
powerline-status==2.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-engineio==3.5.1
python-socketio==3.1.2
requests==2.17.1
six==1.12.0
SQLAlchemy==1.3.2
urllib3==1.21.1
webassets==0.12.1
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.2.1
